Cardano.Prelude.Error

Description

Passing through errors to external libraries that use MonadFail

Documentation

toAesonError ∷ Buildable e ⇒ Either e a → Parser a

Convert an Either-encoded error to an aeson parser error

aesonError ∷ Buildable e ⇒ e → Parser a

Convert a Buildable error into an aeson parser error

toCborError ∷ Buildable e ⇒ Either e a → Decoder s a

Convert an Either-encoded failure to a cborg decoder failure

cborError ∷ Buildable e ⇒ e → Decoder s a

Convert a Buildable error into a cborg decoder error

wrapError ∷ MonadError e' m ⇒ Either e a → (e → e') → m a infix 1

A helper for lifting an Either to a MonadError

By using this function infix we can move the error handling to the end of an expression, hopefully improving readability.

orThrowError ∷ MonadError e m ⇒ Bool → e → m () infix 1

A helper for lifting unless to MonadError

By using this function infix we can move error handling to the end of a Bool expression, hopefully improving readability.