Cardano.Prelude.Formatting

Description

This module contains custom Formats for use with the formatting library

Formats allow a type-safe version of string formatting. For example we might have:

myFormat :: Format r (Text -> Int -> r)
myFormat = "Person's name is " . text . ", age is " . hex

The type parameter r is the polymorphic return type. formatting provides a number of functions to run the format. For example, we could use sformat :: Format Text a -> a to run myFormat, which would give:

sformat myFormat :: Text -> Int -> Text

This is now a simple function, which will return a formatted strict Text.

formatting also provides a Buildable a type-class for values that can be turned into a Text Builder. It provides build :: a -> Builder. There is then a formatter build :: Buildable a => Format r (a -> r). So, to get a Text from any Buildable value we can simply call sformat build value, and we can compose this build in larger Formats.

Documentation

base16Builder ∷ ByteString → Builder

A Builder for a ByteString that performs base 16 encoding

base16F ∷ Format r (ByteString → r)

A Format for a ByteString that performs base 16 encoding

pairF ∷ (Buildable a, Buildable b) ⇒ Format r ((a, b) → r)

A Format for a pair of Buildable values (a, b)

pairBuilder ∷ (Buildable a, Buildable b) ⇒ (a, b) → Builder

A Builder for a pair of Buildable values (a, b)

listJson ∷ (Foldable t, Buildable a) ⇒ Format r (t a → r)

A Format for Foldable containers that outputs a JSON-style list

listJsonIndent ∷ (Foldable t, Buildable a) ⇒ Word → Format r (t a → r)

A Format similar to listJson that prints each value on a new line with indent spaces of indentation

mapJson ∷ (IsList t, Item t ~ (k, v), Buildable k, Buildable v) ⇒ Format r (t → r)

A Format for Exts.IsList containers of Buildable key-value pairs that outputs a JSON-style colon-separated map