Cardano.Prelude.GHC.Heap.NormalForm.Classy

Check a value for unexpected thunks

class NoUnexpectedThunks a where

Check a value for unexpected thunks

Minimal complete definition

Nothing

Methods

noUnexpectedThunks ∷ HasCallStack ⇒ [String] → a → IO ThunkInfo

Check if the argument does not contain any unexpected thunks

For most datatypes, we should have that

noUnexpectedThunks ctxt x == NoUnexpectedThunks

if and only if

isNormalForm x

For some datatypes however, some thunks are expected. For example, the internal fingertree Sequence might contain thunks (this is important for the asymptotic complexity of this data structure). However, we should still check that the values in the sequence don't contain any unexpected thunks.

This means that we need to traverse the sequence, which might force some of the thunks in the tree. In general, it is acceptable for noUnexpectedThunks to force such "expected thunks", as long as it always reports the unexpected thunks.

The default implementation of noUnexpectedThunks checks that the argument is in WHNF, and if so, adds the type into the context (using showTypeOf), and calls whnfNoUnexpectedThunks. See UnexpectedThunkInfo for a detailed discussion of the type context.

See also discussion of caveats listed for isNormalForm.

whnfNoUnexpectedThunks ∷ [String] → a → IO ThunkInfo

Check that the argument is in normal form, assuming it is in WHNF.

The context will already have been extended with the type we're looking at, so all that's left is to look at the thunks inside the type. The default implementation uses GHC Generics to do this.

whnfNoUnexpectedThunks ∷ (Generic a, GWhnfNoUnexpectedThunks '[] (Rep a)) ⇒ [String] → a → IO ThunkInfo

Check that the argument is in normal form, assuming it is in WHNF.

The context will already have been extended with the type we're looking at, so all that's left is to look at the thunks inside the type. The default implementation uses GHC Generics to do this.

showTypeOf ∷ Proxy a → String

Show type a (to add to the context)

We try hard to avoid Typeable constraints in this module: there are types with no Typeable instance but with a NoUnexpectedThunks instance (most important example are types such as ST s which rely on parametric polymorphism). By default we should therefore only show the "outer layer"; for example, if we have a type

Seq (ST s ())

then showTypeOf should just give Seq, leaving it up to the instance for ST to decide how to implement showTypeOf; this keeps things compositional. The default implementation does precisely this using the metadata that GHC Generics provides.

For convenience, however, some of the deriving via newtype wrappers we provide do depend on Typeable; see below.

showTypeOf ∷ (Generic a, GShowTypeOf (Rep a)) ⇒ Proxy a → String

Show type a (to add to the context)

We try hard to avoid Typeable constraints in this module: there are types with no Typeable instance but with a NoUnexpectedThunks instance (most important example are types such as ST s which rely on parametric polymorphism). By default we should therefore only show the "outer layer"; for example, if we have a type

Seq (ST s ())

then showTypeOf should just give Seq, leaving it up to the instance for ST to decide how to implement showTypeOf; this keeps things compositional. The default implementation does precisely this using the metadata that GHC Generics provides.

For convenience, however, some of the deriving via newtype wrappers we provide do depend on Typeable; see below.

allNoUnexpectedThunks ∷ [IO ThunkInfo] → IO ThunkInfo

Short-circuit a list of checks

noUnexpectedThunksInValues ∷ NoUnexpectedThunks a ⇒ [String] → [a] → IO ThunkInfo

Check that all elements in the list are thunk-free

Does not check the list itself. Useful for checking the elements of a container.

See also noUnexpectedThunksInKeysAndValues

noUnexpectedThunksInKeysAndValues ∷ (NoUnexpectedThunks k, NoUnexpectedThunks v) ⇒ [String] → [(k, v)] → IO ThunkInfo

Variant on noUnexpectedThunksInValues for keyed containers.

Neither the list nor the tuples are checked for thunks.

unsafeNoUnexpectedThunks ∷ NoUnexpectedThunks a ⇒ a → Maybe String

Results of the check

data ThunkInfo

Constructors

NoUnexpectedThunks
We found no unexpected thunks

UnexpectedThunk UnexpectedThunkInfo
We found unexpected thunks

data UnexpectedThunkInfo

Information about unexpected thunks

TODO: The ghc-debug work by Matthew Pickering includes some work that allows to get source spans from closures. If we could take advantage of that, we could not only show the type of the unexpected thunk, but also where it got allocated.

Constructors

UnexpectedThunkInfo

Fields

unexpectedThunkContext ∷ [String]

unexpectedThunkCallStack ∷ CallStack

CallStack to where we found the unexpected thunk

NOTE: This is not the callstack of the creation of the thunk.

unexpectedThunkClosure ∷ Maybe Text

The specific closure where we found the problem

This will be unavailable when we used isNormalForm to do the check, since in that case we don't have detailed information available about exactly which thunk was the problem. To aid debugging in such cases buildAndRenderClosureTree can be used to render the closure tree.

thunkInfoToIsNF ∷ ThunkInfo → Bool

Was the value in normal form?

Deriving-via wrappers

newtype UseIsNormalForm a

Newtype wrapper for use with deriving via to use isNormalForm

NoUnexpectedThunks instances derived using UseIsNormalForm will use isNormalForm to check whether the argument is in normal form or not. Since isNormalForm does not need any additional type class instances, this is useful for types that contain fields for which NoUnexpectedThunks instances are not available.

Since the primary use case for UseIsNormalForm then is to give instances for NoUnexpectedThunks from third party libraries, we also don't want to rely on a Generic instance, which may likewise not be available. Instead, we will rely on Typeable, which is available for all types. However, as showTypeOf explains, requiring Typeable may not always be suitable; if it isn't, UseIsNormalFormNamed can be used.

Example:

deriving via UseIsNormalForm T instance NoUnexpectedThunks T

Constructors

UseIsNormalForm a

newtype UseIsNormalFormNamed (name ∷ Symbol) a

Newtype wrapper for use with deriving via to use isNormalForm

This is a variant on UseIsNormalForm which does not depend on Typeable, but instead requires a name to be explicitly given. Example:

deriving via UseIsNormalFormNamed "T" T instance NoUnexpecedThunks T

Constructors

UseIsNormalFormNamed a

newtype OnlyCheckIsWHNF (name ∷ Symbol) a

Newtype wrapper for use with deriving via to check for WHNF only

For some types we don't want to check for nested thunks, and we only want check if the argument is in WHNF, not in NF. A typical example are functions; see the instance of (a -> b) for detailed discussion. This should be used sparingly.

Example:

deriving via OnlyCheckIsWHNF "T" T instance NoUnexpectedThunks T

Constructors

OnlyCheckIsWHNF a

newtype AllowThunk a

Newtype wrapper for values that should be allowed to be a thunk

This should be used VERY sparingly, and should ONLY be used on values (or, even rarer, types) which you are SURE cannot retain any data that they shouldn't. Bear in mind allowing a value of type T to be a thunk might cause a value of type S to be retained if T was computed from S.

Constructors

AllowThunk a

newtype AllowThunksIn (fields ∷ [Symbol]) a

Newtype wrapper for records where some of the fields are allowed to be thunks.

Example: > deriving via AllowThunksIn '["foo","bar"] T instance NoUnexpectedThunks T

This will create an instance that skips the thunk checks for the "foo" and "bar" fields.

Constructors

AllowThunksIn a