Cardano.Prelude.GHC.Heap.Size

Documentation

data CountFailure

Constructors

WorkListFull

VisitedFull

OutOfMemory

UnsupportedClosure ClosureType

data PerformGC

Should we perform a GC call before counting the size?

Constructors

FirstPerformGC
Yes, first perform GC before counting

This should be used for most accurate results. Without calling GC first, the computed size might be larger than expected due to leftover indirections (black holes, selector thunks, etc.)

DontPerformGC
No, do not perform GC before counting

If pinpoint accuracy is not requried, then GC can be skipped, making the call much less expensive.

computeHeapSize ∷ a → IO (Either CountFailure Word64)

Compute the size of the given closure

This is a wrapper around computeHeapSize' which sets some defaults for the capacity of worklist and the visited set: it uses a worklist capacity of 10k (which, assuming balanced data structures, should be more than enough), an initial visited set capacity of 250k, and a maximum visited set capacity of 16M. This means that this will use between 2 MB and 128 MB of heaps space.

It also does NOT perform GC before counting, for improved performance. Client code can call performMajorGC manually or use computeHeapSize'.

Should these limits not be sufficient, or conversely, the memory requirements be too large, use computeHeapSize' directly.

computeHeapSize'

Arguments

∷ PerformGC
Should we call GC before counting?

→ Word
Capacity of the worklist

→ Word
Initial capacity of the visited set

→ Word
Maximum capacity of the visited set

→ a

→ IO (Either CountFailure Word64)

Compute the size of the given closure

The size of the worklist should be set to the maximum expected depth of the closure; the size of the visited set should be set to the maximum /number of nodes/ in the closure.

computeHeapSizeWorkList can be used to estimate the size of the worklist required.

computeHeapSizeWorkList ∷ a → Word64

Upper bound on the required work list size to compute closure size

NOTE: This ignores sharing, and so provides an upper bound only.

The size of a closure with no nested pointers can be computed without any stack space.

When we have a closure with (N + 1) nested pointers

p0 p1 .. pN

We will

Push pN, .., p1, p0 onto the stack

Pop off p0 and count its children

Pop off p1 and count its children

..

until we have processed all children. This means that the stack space required will be the maximum of

[ N + 1 -- For the initial list
, requiredWorkList p0 + (N + 1) - 1
, requiredWorkList p1 + (N + 1) - 2
, ..
, requiredWorkList pN + (N + 1) - (N + 1)
]

For example, for a list, we would get that

requiredWorkList []     == 0
requiredWorkList (x:xs) == max [ 2
                               , requiredWorkList x + 1
                               , requiredWorkList xs
                               ]

which, for a list of Int (which requires only a stack of size 1), equals 2 (unless the list is empty).

Similarly, for binary trees, we get

requiredWorkList Leaf           == 0
requiredWorkList (Branch l x r) == max [ 3
                                       , requiredWorkList l + 2
                                       , requiredWorkList x + 1
                                       , requiredWorkList r
                                       ]

which, for a tree of Int, is bound by (height * 2) + 1.