Cardano.Prelude.GHC.Heap.Tree

Description

This module exports functions and data types for building and printing GHC heap object trees. Very useful for debugging issues involving the heap representation of Haskell expressions.

Documentation

data ClosureTreeOptions

Options which detail how a Closure Tree should be constructed.

Constructors

ClosureTreeOptions

Fields

ctoMaxDepth ∷ !TreeDepth

Construct a closure tree given a maximum depth.

ctoCyclicClosures ∷ !TraverseCyclicClosures

Whether to traverse cyclic closures while constructing a closure tree.

data TraverseCyclicClosures

Whether to traverse cyclic closures in a Closure Tree.

Constructors

TraverseCyclicClosures
Traverse cyclic closures.

NoTraverseCyclicClosures
Do not traverse cyclic closures.

data TreeDepth

The depth of a Tree.

Constructors

TreeDepth !Int
A specific tree depth bound.

AnyDepth
No tree depth bound.

buildClosureTree ∷ ClosureTreeOptions → a → IO (Maybe (Tree Closure))

Given a Haskell expression, build a Tree which reflects its heap object representation.

buildAndRenderClosureTree ∷ ClosureTreeOptions → a → IO Text

Given a Haskell expression, build a Tree which reflects its heap object representation and render it as Text.

depth ∷ Tree a → Int

isZeroOrNegativeTreeDepth ∷ TreeDepth → Bool

renderClosure ∷ Closure → Text

renderTree ∷ Tree a → (a → Text) → Text