Cardano.Prelude.Show

Description

This module uses GHC.Generics to define gshowsPrecWithoutRecordSyntax, which can often make the result of show more legible, especially for newtypes for example.

Documentation

gshowsPrecWithoutRecordSyntax ∷ ∀ a. (Generic a, Coercible (Rep a) (WithoutRecordSyntax (Rep a)), GShow' (WithoutRecordSyntax (Rep a))) ⇒ Int → a → ShowS

Implements showsPrec as if the constructors' record fields were instead plain arguments

newtype GShowK1 i c p

Merely an implementation detail

This constructor must be in scope at usage sites of gshowsPrecWithoutRecordSyntax for its Coercible constraint to be satisfiable.

Constructors

GShowK1 (K1 i c p)

type family WithoutRecordSyntax (rep ∷ Type → Type) ∷ Type → Type where ...

An implementation detail of gshowsPrecWithoutRecordSyntax

This removes the metadata in a GHC.Generics Rep instance incurred by the use of record syntax in the data type declaration. It also replaces K1 with GShowK1.

The result is as if the constructors' record fields were instead plain arguments, modulo the K1 and GShowK1 swap.

Equations

WithoutRecordSyntax (M1 i c f) = M1 i (WithoutRecordSyntaxMeta c) (WithoutRecordSyntax f)

WithoutRecordSyntax (K1 i c) = GShowK1 i c

WithoutRecordSyntax (a :+: b) = WithoutRecordSyntax a :+: WithoutRecordSyntax b

WithoutRecordSyntax (a :*: b) = WithoutRecordSyntax a :*: WithoutRecordSyntax b

WithoutRecordSyntax (a :.: b) = WithoutRecordSyntax a :.: WithoutRecordSyntax b

WithoutRecordSyntax V1 = V1

WithoutRecordSyntax U1 = U1

WithoutRecordSyntax UChar = UChar

WithoutRecordSyntax UDouble = UDouble

WithoutRecordSyntax UFloat = UFloat

WithoutRecordSyntax UInt = UInt

WithoutRecordSyntax UWord = UWord