Cardano.Prelude.Strict

Description

Helper functions for enforcing strictness.

Documentation

forceElemsToWHNF ∷ Foldable t ⇒ t a → t a

Force all of the elements of a Foldable to weak head normal form.

In order to ensure that all of the elements of a Foldable are strict, we can simply foldMap over it and seq each value with '()'. However, '()''s mappend implementation is actually completely lazy: _ <> _ = () So, in order to work around this, we instead utilize this newly defined StrictUnit whose mappend implementation is specifically strict.